Refactor: Centralize clean_markdown_code_blocks in utils and update imports to absolute paths
