+++ src/squirrel/dates.py
@@ -6,6 +6,7 @@ from langchain_core.prompts import ChatPromptTemplate
 from langchain_core.runnables import Runnable
 
 from squirrel.prompts import CleaningStrOutputParser, llm_engine
+from squirrel.utils import clean_markdown_code_blocks
 
 
 def get_current_date_context() -> str:
@@ -85,36 +86,6 @@ def date_processing_chain() -> Runnable:
     )
 
     return prompt | llm_engine | CleaningStrOutputParser()
-
-
-def clean_markdown_code_blocks(text: str) -> str:
-    """
-    Remove markdown code blocks from text regardless of language.
-
-    Args:
-        text: Text that may contain markdown code blocks
-
-    Returns:
-        Cleaned text without markdown code blocks
-    """
-    text = text.strip()
-
-    # Check if text starts with ``` followed by optional language identifier
-    if text.startswith("```"):
-        # Find the first newline after the opening ```
-        first_newline = text.find("\n")
-        if first_newline != -1:
-            # Remove everything up to and including the first newline
-            text = text[first_newline + 1 :]
-        else:
-            # No newline found, remove the opening ```
-            text = text[3:]
-
-    # Remove closing ``` if present
-    if text.endswith("```"):
-        text = text[:-3]
-
-    return text.strip()
 
 
 def process_question_dates(question: str) -> Dict[str, Any]:
